Undo second Flask Blueprint and finish basic API-related functions
Realized there was really no need for second Blueprint since API-related functions can simply be imported into the views.py file which will keeps things simpler and cleaner. Added a bit of logic for testing user access dependent on the session and a new HTML page for when the user passes 'login'.