Add self-made decorator to protect proutes from users not logged in
With the MVP, Strava produced the error page for users not logged in.
With the MVP+1, the JS Leaflet map was rendering but without activities so a self-made protective layer was necessary to prevent this. Because a literal 'log in' does not exist (it has essentially been offloaded to Strava), the Flask built-in did not seem suitable. The self-made decorator checks if the user in is the session and, if not, redirects to index.html with a flash prompting authentication via Strava to see the other pages.